draft(member-cards): WIP
