+++ app/fragments/app-member-cards-grid/src/hooks/check-members-onbording-conditions.hook.ts
@@ -1,23 +1,17 @@
-// import { getDiscordServerMembers } from '@globals/third-party-service-data'
-
 import { INIT_ONBORDING_DATA } from '../member-cards-grid.constants.js'
 
 export const checkMembersOnbordingConditions = async ({
   membersData,
-  organizationsData,
   membersInDefaultOrganization,
   membersWithou2fa,
 }) => {
-  // const discordServerMembers = await getDiscordServerMembers()
-  // console.log(discordServerMembers)
-
   const resultMembersData = await membersData.map((memberData) => {
-    // console.log(memberData)
-    const { avatarUrl, id: memberId } = memberData
+    const { id: memberId } = memberData
 
-    const memberWithOnbordingData = memberData
-
-    memberWithOnbordingData.onbordingData = [...INIT_ONBORDING_DATA]
+    let memberWithOnbordingData = {
+      ...memberData,
+      onbordingData: [...INIT_ONBORDING_DATA],
+    }
 
     if (membersInDefaultOrganization.includes(memberId)) {
       memberWithOnbordingData.onbordingData.forEach((conditionData, conditionIndex) => {
