+++ app/fragments/app-member-cards-grid/src/hooks/helpers/check-member-in-default-organization.helper.ts
@@ -1,0 +1,1 @@
+export const checkMemberInDefaultOrganization = () => {}
